+++ src/QueryType/ManagedResources/Query/Command/Exists.php
@@ -19,11 +19,11 @@ use Solarium\QueryType\ManagedResources\Query\AbstractQuery as Query;
 class Exists extends AbstractCommand
 {
     /**
-     * Term to be checked if exists.
+     * Name of the child resource to be checked if exists.
      *
-     * @var string
+     * @var string|null
      */
-    protected $term = '';
+    protected $term = null;
 
     /**
      * Returns command type, for use in adapters.
@@ -42,31 +42,28 @@ class Exists extends AbstractCommand
      */
     public function getRequestMethod(): string
     {
-        return Request::METHOD_GET;
+        $method = Request::METHOD_HEAD;
+
+        // there's a bug since Solr 8.7 with HEAD requests if a term is set (SOLR-15116)
+        if (null !== $this->getTerm()) {
+            $method = Request::METHOD_GET;
+        }
+
+        return $method;
     }
 
     /**
-     * Empty.
+     * Returns the name of the child resource to be checked if exists.
      *
-     * @return string
+     * @return string|null
      */
-    public function getRawData(): string
-    {
-        return '';
-    }
-
-    /**
-     * Returns the term to be checked if exists.
-     *
-     * @return string
-     */
-    public function getTerm(): string
+    public function getTerm(): ?string
     {
         return $this->term;
     }
 
     /**
-     * Set the term to be checked if exists.
+     * Set the name of the child resource to be checked if exists.
      *
      * @param string $term
      *
@@ -75,6 +72,18 @@ class Exists extends AbstractCommand
     public function setTerm(string $term): self
     {
         $this->term = $term;
+
+        return $this;
+    }
+
+    /**
+     * Remove the name of the child resource. This reverts to checking if the managed resource exists.
+     *
+     * @return self
+     */
+    public function removeTerm(): self
+    {
+        $this->term = null;
 
         return $this;
     }